Fixing patch test and some stuff around model policies
